Update binary practice
Athen: This is what I have figured out and it seems to work. I have not experimented with many file types, so you should see if this works for a wide variety of them. Also, it would be good to rewrite the encryption and decryption functions so that it is not just a shift cipher.